+++ src/routes/users.js
@@ -1,0 +1,13 @@
+const path = require('path');
+const express = require('express');
+const router = express.Router();
+
+router.get('/register', (req, res) => {
+    res.sendFile(path.join(__dirname, '../views/register.html'));
+});
+
+router.get('/login', (req, res) => {
+    res.sendFile(path.join(__dirname, '../views/login.html'));
+});
+
+module.exports = router;
