+++ src/views/index.html
@@ -27,7 +27,7 @@
             <nav class="header-navbar">
                 <ul>
                     <li><a href="/products">Comprar</a></li>
-                    <li><a href="/login"><i class="fas fa-user-circle"></i>Perfil</a></li>
+                    <li><a href="/user/login"><i class="fas fa-user-circle"></i>Perfil</a></li>
                     <li><a href="">Favoritos</a></li>
                     <li><a href="/cart"><i class="fas fa-shopping-cart"></i></a></li>
                 </ul>
@@ -42,7 +42,7 @@
             </div>
             <h2>Autos Destacados</h2>
             <div class="featured-cars">
-                <a href="/product-detail">
+                <a href="/products/product-detail">
                     <article>
                         <img src="images/ford-fiesta.jpeg" alt="ford-fiesta">
                         <h3>Ford Fiesta</h3>
