+++ src/views/productList.html
@@ -48,7 +48,7 @@
                 </ul>
             </div>
             <div class="productos">
-                    <a href="/product-detail">
+                    <a href="/products/product-detail">
                         <article>
                             <img src="images/ford-fiesta.jpeg" alt="ford-fiesta">
                             <h3>Ford Fiesta</h3>
